src/main/java/com/example/demo/filter/security/JwtTokenFilter.java
package com.example.demo.filter.security;

import com.example.demo.data.security.JwtToken;
import com.example.demo.service.security.JwtService;
import io.jsonwebtoken.JwtException;
import jakarta.servlet.FilterChain;
import jakarta.servlet.ServletException;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import org.springframework.http.HttpStatus;
import org.springframework.lang.NonNull;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
import org.springframework.stereotype.Component;
import org.springframework.util.StringUtils;
import org.springframework.web.filter.OncePerRequestFilter;

import java.io.IOException;

import static com.example.demo.utils.security.SecurityConstants.*;

@Component
public class JwtTokenFilter extends OncePerRequestFilter {

    private final JwtService jwtService;
    private final UserDetailsService userDetailsService;

    public JwtTokenFilter(JwtService jwtService, UserDetailsService userDetailsService) {
        this.jwtService = jwtService;
        this.userDetailsService = userDetailsService;
    }

    @Override
    protected void doFilterInternal(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response,
                                    @NonNull FilterChain filterChain) throws ServletException, IOException {
        final String authHeader = request.getHeader(AUTHORIZATION_HEADER);
        if (!StringUtils.hasLength(authHeader) || !authHeader.startsWith(BEARER_AUTHORIZATION)) {
            filterChain.doFilter(request, response);
            return;
        }

        final String jwt = authHeader.substring(BEARER_AUTHORIZATION.length());

        UserDetails userDetails;
        try {
            final JwtToken token = jwtService.validateAndGetToken(jwt);

            if ((token.isRefreshToken() && !request.getRequestURI().endsWith(TOKEN_REFRESH_ENDPOINT))
                    || (token.isAccessToken() && request.getRequestURI().endsWith(TOKEN_REFRESH_ENDPOINT))) {
                throw new JwtException(INVALID_TOKEN_TYPE);
            }

            userDetails = userDetailsService.loadUserByUsername(token.subject());
        } catch (JwtException | UsernameNotFoundException ex) {
            response.setStatus(HttpStatus.UNAUTHORIZED.value());
            return;
        }

        if (SecurityContextHolder.getContext().getAuthentication() == null) {
            UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
                    userDetails,
                    null,
                    userDetails.getAuthorities()
            );

            authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
            SecurityContextHolder.getContext().setAuthentication(authToken);
        }

        filterChain.doFilter(request, response);
    }

    @Override
    protected boolean shouldNotFilterErrorDispatch() {
        return false;
    }
}
